requirements.txt
notion-client
pandas
matplotlib
python-dotenv
python-dotenv
wordcloud
